feat: /meets/home API 및 Response 정렬
