--- src/main/java/com/swyp/mema/domain/locaction/dto/response/SingleLocationResponse.java
@@ -1,12 +1,0 @@
-package com.swyp.mema.domain.locaction.dto.response;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class SingleLocationResponse {
-
-	private String startStation;
-
-}
