--- src/main/java/com/swyp/mema/domain/locaction/dto/response/TotalLocationResponse.java
@@ -1,14 +1,0 @@
-package com.swyp.mema.domain.locaction.dto.response;
-
-import java.util.List;
-
-import lombok.Builder;
-import lombok.Getter;
-
-@Getter
-@Builder
-public class TotalLocationResponse {
-
-	private List<String> startStationList;
-	private String arrivalStation;
-}
