+++ src/main/java/com/swyp/mema/domain/meetMember/repository/MeetMemberCustomRepository.java
@@ -2,9 +2,14 @@ package com.swyp.mema.domain.meetMember.repository;
 
 import java.util.List;
 
+import com.swyp.mema.domain.meet.dto.response.MeetHomeDetailResponse;
+import com.swyp.mema.domain.meet.model.Meet;
 import com.swyp.mema.domain.meetMember.dto.response.MeetMemberRes;
 import com.swyp.mema.domain.user.dto.response.UserRes;
 
 public interface MeetMemberCustomRepository {
+
 	List<MeetMemberRes> findMeetMembersWithUserInfo(Long meetId); // 약속원과 사용자 정보 조회
+
+	List<Meet> findMeetsByUserId(Long userId);
 }
