+++ src/main/java/com/swyp/mema/domain/meetMember/repository/MeetMemberCustomRepositoryImpl.java
@@ -4,6 +4,8 @@ import java.util.List;
 
 import com.querydsl.core.types.Projections;
 import com.querydsl.jpa.impl.JPAQueryFactory;
+import com.swyp.mema.domain.meet.model.Meet;
+import com.swyp.mema.domain.meet.model.QMeet;
 import com.swyp.mema.domain.meetMember.dto.response.MeetMemberRes;
 import com.swyp.mema.domain.meetMember.model.QMeetMember;
 import com.swyp.mema.domain.user.dto.response.UserRes;
@@ -39,6 +41,21 @@ public class MeetMemberCustomRepositoryImpl implements MeetMemberCustomRepositor
 			.from(qMeetMember)
 			.join(qUser).on(qMeetMember.user.userId.eq(qUser.userId))
 			.where(qMeetMember.meet.id.eq(meetId))
+			.fetch();
+	}
+
+	@Override
+	public List<Meet> findMeetsByUserId(Long userId) {
+
+		QMeetMember meetMember = QMeetMember.meetMember;
+		QMeet meet = QMeet.meet;
+
+		return queryFactory
+			.select(meet)
+			.from(meetMember)
+			.join(meetMember.meet, meet)
+			.where(meetMember.user.userId.eq(userId)) // userId 조건
+			.distinct() // 중복 제거
 			.fetch();
 	}
 }
