+++ src/main/java/com/swyp/mema/domain/meetMember/repository/MeetMemberRepository.java
@@ -1,5 +1,6 @@
 package com.swyp.mema.domain.meetMember.repository;
 
+import java.util.List;
 import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
