+++ src/main/java/com/swyp/mema/domain/station/repository/StationRepository.java
@@ -1,0 +1,8 @@
+package com.swyp.mema.domain.station.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import com.swyp.mema.domain.station.model.Station;
+
+public interface StationRepository extends JpaRepository<Station, String> {
+}
