+++ src/main/java/com/swyp/mema/domain/user/dto/response/UserRes.java
@@ -2,10 +2,12 @@ package com.swyp.mema.domain.user.dto.response;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.AllArgsConstructor;
+import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 @Getter
+@Builder
 @NoArgsConstructor
 @AllArgsConstructor
 @Schema(description = "유저 프로필 응답")
