+++ src/main/java/com/swyp/mema/domain/voteLocation/dto/request/CreateLocationReq.java
@@ -1,0 +1,19 @@
+package com.swyp.mema.domain.voteLocation.dto.request;
+
+import jakarta.validation.constraints.NotBlank;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public class CreateLocationReq {
+
+	@NotBlank
+	private String stationId;
+
+	@NotBlank
+	private String stationName;
+
+	@NotBlank
+	private String routeName;
+}
