+++ src/main/java/com/swyp/mema/domain/voteLocation/dto/response/SingleLocationResponse.java
@@ -1,0 +1,12 @@
+package com.swyp.mema.domain.voteLocation.dto.response;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public class SingleLocationResponse {
+
+	private String startStation;
+
+}
