+++ src/main/java/com/swyp/mema/global/base/domain/BaseEntity.java
@@ -27,8 +27,5 @@ public abstract class BaseEntity {
 	@LastModifiedDate
 	@Column(name = "update_date", nullable = false)
 	private LocalDateTime updateDate;
-
-	@Column(name = "delete_yn", nullable = false, columnDefinition = "BOOLEAN DEFAULT FALSE")
-	private boolean deleteYn = false;
 }
